requirements.txt
wheel
setuptools
gunicorn
httptools
Eel==0.12.4
fastapi==0.63.0
fasteners==0.14.1
finta==1.3
future==0.18.2
h11==0.9.0
hidapi==0.10.1
hijri-converter==2.1.1
holidays==0.11.1
html5lib==1.0.1
httplib2==0.14.0

idna==2.8
matplotlib==3.3.4
openpyxl==3.0.6
python-multipart==0.0.5
pyspectra
PyInstaller
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.24.1
scipy==1.5.4
simplejson==3.16.0
six==1.14.0
sklearn==0.0
urllib3==1.25.8
uvicorn==0.11.3
wadllib==1.3.3
wcwidth==0.1.8
webargs
webencodings==0.5.1
websockets==8.1
wrapt==1.11.2
